Auth/pyproject.toml
[project]
name = "template-repo"
version = "0.1.0"
description = ""
authors = [
    {name = "csejh1",email = "jerrysong4912@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "prometheus-fastapi-instrumentator (>=7.1.0,<8.0.0)",
    "aioboto3 (>=15.5.0,<16.0.0)",
    "aiomysql (>=0.3.2,<0.4.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "pymysql (>=1.1.2,<2.0.0)",
    "cryptography (>=46.0.3,<47.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.14.10,<0.15.0)",
    "poethepoet (>=0.39.0,<0.40.0)"
]

[tool.poetry]
package-mode = false

[tool.poe.tasks.copy-env]
# script 대신 shell을 사용하여 파이썬 한 줄 명령어를 실행합니다.
shell = "python -c \"import shutil, os; os.path.exists('.env') or shutil.copy('.env.example', '.env')\""

[tool.poe.tasks]
setup = "poetry install --no-root"
run   = "uvicorn app.main:app --reload"
lint  = [
    { cmd = "ruff check --fix ." },
    { cmd = "ruff format ." }
]
# --- DB 관련 명령어 ---
db-up   = "docker compose up -d"
db-down = "docker compose down"
db-clean = "docker compose down -v"
migrate = "alembic upgrade head"
# 여기에 인자를 전달받을 수 있도록 설정되어 있습니다.
makemigrations = "alembic revision --autogenerate"